sysctl: Move sysctl package under datapath/linux
Sysctl should be part of the datapath module, so the commit moves the
package accordingly.
