+++ pkg/datapath/connector/add.go
@@ -9,9 +9,9 @@ import (
 
 	"golang.org/x/sys/unix"
 
+	"github.com/cilium/cilium/pkg/datapath/linux/sysctl"
 	"github.com/cilium/cilium/pkg/logging"
 	"github.com/cilium/cilium/pkg/logging/logfields"
-	"github.com/cilium/cilium/pkg/sysctl"
 )
 
 var log = logging.DefaultLogger.WithField(logfields.LogSubsys, "endpoint-connector")
